resolved the comments
