+++ README.md
@@ -5,3 +5,13 @@ First, run the development server:
 ```bash
 npm run dev
 ```
+Create an .env file and add these variables:
+
+MEETUP_API_BASE_URL=https://www.meetup.com
+
+SETTINGS_BASE_URL=https://raw.githubusercontent.com/CodeandCoffeeCommunity/Code-and-Coffee-Website-Service-Settings/production/settings
+
+
+Reach out to Website Team to get these keys
+NOTION_API_KEY = secret...
+NOTION_API_DATABASE =
