+++ app/(routes)/organizers/page.tsx
@@ -1,33 +1,11 @@
+
 import "@/app/globals.css";
 import Wrapper from "@/components/layout/Wrapper";
-import { Client } from "@notionhq/client";
 
-const notion = new Client({ auth: process.env.NOTION_API_KEY });
+//import the function that fetches data from Notion
+import fetchDataFromNotion from "@/app/service/notion/NotionService"
 
-// this function fetches data from The Notion Database(Table). 
-const fetchDataFromNotion = async () => {
-  const databaseId = process.env.NOTION_API_DATABASE;
-  const response = await notion.databases.query({ database_id: databaseId });
 
-  // we map through the array of object to pull out the properties that we need for rendering.
-  const teamResults = response.results.map((page) => {
-    return {
-      id: page.id,
-      name: page.properties.Name.title[0]?.plain_text,
-      role: page.properties.Role.multi_select[0].name,
-      // team: page.properties.Teams.rich_text[0]?.plain_text,
-      imageUrl: page.properties.Image.files[0]?.file.url,
-      linkedInUrl: page.properties.LinkedIn.rich_text[0]?.plain_text,
-      notionId: page.properties.NotionId.rich_text[0]?.plain_text
-    };
-  });
-
-  //this function sorts the array by NotionId property that we set in the Notion table
-  teamResults.sort((a, b) => {
-    return parseInt(a.notionId) - parseInt(b.notionId);
-  });
-  return teamResults;
-};
 
 
 export default async function Organizers() {
