+++ package.json
@@ -23,7 +23,6 @@
     "dotenv": "^16.3.1",
     "eslint-config-next": "13.4.3",
     "next": "13.4.3",
-    "node-fetch": "^3.3.2",
     "postcss": "8.4.23",
     "react": "18.2.0",
     "react-dom": "18.2.0",
